Minor refactoring of authenication.
